Moved different distribution logic to different modules
